refactor fragments
