+++ src/morphdom/constants.js
@@ -1,0 +1,4 @@
+exports.___ELEMENT_NODE = 1;
+exports.___TEXT_NODE = 3;
+exports.___COMMENT_NODE = 8;
+exports.___COMPONENT_NODE = 2;
